refactor(tools): extract search domain into src/tools/search.py
